fix when a user changes his type user to professor

+++ app/controllers/users/registrations_controller.rb
@@ -1,6 +1,6 @@
 class Users::RegistrationsController < Devise::RegistrationsController
-before_action :configure_sign_up_params, only: [:create]
-before_action :configure_account_update_params, only: [:update]
+  before_action :configure_sign_up_params, only: [:create]
+  before_action :configure_account_update_params, only: [:update]
 
   # GET /resource/sign_up
   # def new
@@ -29,20 +29,26 @@ before_action :configure_account_update_params, only: [:update]
     if resource_updated
 
       if resource.type_user == "professor"
-        list_students = params[:user][:users_advisor].map{|x| x[1]}
-        list_students_temp = UsersAdvisor.where(:professor_id => resource.id).map(&:student_id)
+        list_students = params[:user][:users_advisor]
 
-        if list_students.empty?
-          UsersAdvisor.where(:professor_id => resource.id).each do |u|
-            u.delete
-          end
-        else
-          list_to_delete = list_students_temp - list_students
-          list_to_add = list_students - list_students_temp
+        if not list_students.nil?
+          list_students = list_students.map{|x| x[1]}
 
-          UsersAdvisor.where(:professor_id => resource.id, :student_id => list_to_delete).delete_all
-          list_to_add.each do |u|
-            UsersAdvisor.create(:professor_id => resource.id, :student_id => u)
+          list_students_temp = UsersAdvisor.where(:professor_id => resource.id).map(&:student_id)
+
+          if list_students.empty?
+            UsersAdvisor.where(:professor_id => resource.id).each do |u|
+              u.delete
+            end
+          else
+            list_to_delete = list_students_temp - list_students
+            list_to_add = list_students - list_students_temp
+
+            UsersAdvisor.where(:professor_id => resource.id, :student_id => list_to_delete).delete_all
+            list_to_add.each do |u|
+              UsersAdvisor.create(:professor_id => resource.id, :student_id => u)
+            end
+
           end
         end
       end
